requirements.txt
mcp>=1.0.0
pyyaml>=6.0
pymupdf>=1.24.0
